+++ packages/maestro_cli/lib/src/common/paths.dart
@@ -7,9 +7,11 @@ const maestroArtifactPathEnv = 'MAESTRO_ARTIFACT_PATH';
 
 String get serverArtifact => 'server-$version';
 String get instrumentationArtifact => 'instrumentation-$version';
-
 String get serverArtifactFile => '$serverArtifact.apk';
 String get instrumentationArtifactFile => '$instrumentationArtifact.apk';
+
+String get debugServerArtifactFile => 'server.apk';
+String get debugInstrumentationArtifactFile => 'instrumentation.apk';
 
 /// [artifact] must be in the form of `$artifact-$version`, for example:
 /// `server-1.0.0`.
@@ -23,8 +25,16 @@ String get serverArtifactPath {
   return path.join(artifactPath, serverArtifactFile);
 }
 
+String get debugServerArtifactPath {
+  return path.join(artifactPath, debugServerArtifactFile);
+}
+
 String get instrumentationArtifactPath {
   return path.join(artifactPath, instrumentationArtifactFile);
+}
+
+String get debugInstrumentationArtifactPath {
+  return path.join(artifactPath, debugInstrumentationArtifactFile);
 }
 
 bool get artifactPathSetFromEnv {