+++ packages/maestro_test/lib/maestro_test.dart
@@ -3,3 +3,5 @@
 library maestro_test;
 
 export 'src/automator.dart';
+export 'src/conditions.dart';
+export 'src/native_widget.dart';